Add files via upload
This is a system for city planners to automatically review building plans for new homes. The zoning is based on Eagle, Idaho's city code. This is my first custom script, so I recognize that it's not the most efficient thing out there, but it works. Feel free to advise and improve the code.